backend/requirements.txt
fastapi>=0.103.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
pydantic>=1.10.0
scikit-learn>=1.3.0
pandas>=2.1.0
numpy>=1.25.0
joblib>=1.3.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9
